Harden ui input verbs per PR #587 review (#562, #498)
Address the pr-review findings on the send-keys (#562) and drag / scroll
--wheel (#498) input-injection verbs.

High:
- H1 ForegroundGuard: compare the foreground window's root ancestor
  (GetAncestor GA_ROOT) to the target instead of PID-matching, so a
  child/owned window of the target counts as "foreground" and a same-PID
  but different top-level window does not. Drops the unsafe block.
- H2 drag from-point settle race: move the cursor, wait, re-confirm the
  from element is still there (ConfirmStillAsync), re-check the foreground,
  then press with settleMs:0. Mirrors the click fix so "success" means the
  button-down landed on the resolved target, not empty space.

Medium:
- M1 KeyStringParser: a modifier-led chord with an empty segment around
  '+' ("ctrl++a", "ctrl+", "shift++enter") now throws instead of silently
  dropping the empty and pressing a different chord. Non-modifier-led
  tokens ("a+b", "C++", "ctrl+a+b") stay literal text.
- M2 scroll --wheel: same cursor-settle / confirm / foreground sequence as
  drag before synthesizing the wheel.
- M3 click: re-check the foreground immediately before the button-down
  (second gate), closing the same settle-window race on the success path.
- M7 drag: a comma-shaped from/to token that isn't a valid x,y pair
  ("100,", "100,200,300") returns invalid_arguments instead of being
  treated as a selector and misreported as element-not-found.
- M8 --verbatim: whitespace-only is legitimate verbatim content (types the
  spaces) rather than a "nothing to send" error; only genuinely empty is
  rejected.
- M9 send-keys --via send-input: refuse OS-wide injection when no target
  window resolves (keystrokes have no location to verify the foreground
  against) instead of typing blindly into whatever has focus.
- M10 send-keys: the human log reports the action count, not the raw
  keystrokes (which may carry secrets in persisted CI logs); --json still
  includes `keys` for opt-in callers.

Tests:
- M4 split the 1.5k-line UiCommandTests.cs into partial-class files
  (UiCommandTests.SendKeys.cs, UiCommandTests.Mouse.cs) so no file exceeds
  the ~1000-line guideline; behavior-preserving.
- M5 foreground-deny tests for drag, scroll --wheel, and send-keys
  --via send-input.
- M6 drag from-point moves-during-settle test (mirrors the click case).
- Added malformed-chord, verbatim-whitespace, malformed-coordinate, and
  no-resolvable-target tests.

1540 pass / 5 pre-existing E2E_Node env fails / 1 skip. cli-schema content
unchanged (no Description edits).
