+++ src/winapp-CLI/WinApp.Cli.Tests/UiCommandTests.cs
@@ -9,7 +9,7 @@ using WinApp.Cli.Services;
 namespace WinApp.Cli.Tests;
 
 [TestClass]
-public class UiCommandTests : BaseCommandTests
+public partial class UiCommandTests : BaseCommandTests
 {
     private FakeUiAutomationService _fakeUia = null!;
     private FakeUiSessionService _fakeSession = null!;
@@ -816,7 +816,7 @@ public class UiCommandTests : BaseCommandTests
 
         Assert.AreEqual(0, exitCode);
         Assert.AreEqual(1, _fakeMouse.ClickCalls.Count);
-        Assert.AreEqual(1, _fakeForeground.Calls.Count);
+        Assert.AreEqual(2, _fakeForeground.Calls.Count, "click checks the foreground twice on the success path: once before the cursor-settle confirm read and once immediately before the button-down");
         Assert.AreEqual(7777, _fakeForeground.Calls[0].TargetHwnd);
     }
 
@@ -841,547 +841,6 @@ public class UiCommandTests : BaseCommandTests
         var command = GetRequiredService<UiScrollCommand>();
         var exitCode = await ParseAndInvokeWithCaptureAsync(command, [sel, "-a", "TestApp", "--wheel", "-3", "--json"]);
 
-        Assert.AreEqual(1, exitCode);
-        Assert.AreEqual(0, _fakeMouse.ScrollWheelCalls.Count);
-    }
-
-    // ---------------------------------------------------------------------
-    // send-keys (#562) — synthetic keyboard input
-    // ---------------------------------------------------------------------
-
-    [TestMethod]
-    public async Task SendKeys_Json_EmitsEnvelope()
-    {
-        var command = GetRequiredService<UiSendKeysCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command, ["ctrl+a", "-a", "TestApp", "--json"]);
-        Assert.AreEqual(0, exitCode);
-
-        var result = System.Text.Json.JsonSerializer.Deserialize<System.Text.Json.JsonElement>(TestAnsiConsole.Output);
-        Assert.AreEqual("ctrl+a", result.GetProperty("keys").GetString());
-        Assert.AreEqual("post-message", result.GetProperty("via").GetString());
-        Assert.AreEqual(1, result.GetProperty("actionCount").GetInt32());
-    }
-
-    [TestMethod]
-    public async Task SendKeys_DefaultTransport_IsPostMessage()
-    {
-        var command = GetRequiredService<UiSendKeysCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command, ["enter", "-a", "TestApp"]);
-        Assert.AreEqual(0, exitCode);
-
-        Assert.AreEqual(1, _fakeKeyboard.SendCalls.Count);
-        Assert.AreEqual(WinApp.Cli.Helpers.KeyTransport.PostMessage, _fakeKeyboard.SendCalls[0].Transport);
-    }
-
-    [TestMethod]
-    public async Task SendKeys_ViaSendInput_SetsTransport()
-    {
-        var command = GetRequiredService<UiSendKeysCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command, ["enter", "-a", "TestApp", "--via", "send-input"]);
-        Assert.AreEqual(0, exitCode);
-
-        Assert.AreEqual(1, _fakeKeyboard.SendCalls.Count);
-        Assert.AreEqual(WinApp.Cli.Helpers.KeyTransport.SendInput, _fakeKeyboard.SendCalls[0].Transport);
-    }
-
-    [TestMethod]
-    public async Task SendKeys_SequenceAndText_ParsesMultipleActions()
-    {
-        var command = GetRequiredService<UiSendKeysCommand>();
-        // "down down enter" -> 3 key chords; "hello" -> 1 text action
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command, ["down down enter hello", "-a", "TestApp"]);
-        Assert.AreEqual(0, exitCode);
-
-        Assert.AreEqual(1, _fakeKeyboard.SendCalls.Count);
-        var actions = _fakeKeyboard.SendCalls[0].Actions;
-        Assert.AreEqual(4, actions.Count);
-        Assert.IsInstanceOfType<WinApp.Cli.Helpers.TextInput>(actions[3]);
-    }
-
-    [TestMethod]
-    public async Task SendKeys_WithTarget_FocusesElementAndUsesElementHwnd()
-    {
-        _fakeUia.FindSingleResult = new UiElement
-        {
-            Id = "e0", Type = "Edit", Selector = "txt-name-1234", WindowHandle = 4242
-        };
-
-        var command = GetRequiredService<UiSendKeysCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command,
-            ["hello", "-a", "TestApp", "--target", "txt-name-1234", "--json"]);
-        Assert.AreEqual(0, exitCode);
-
-        Assert.AreEqual(1, _fakeKeyboard.SendCalls.Count);
-        Assert.AreEqual(4242, _fakeKeyboard.SendCalls[0].Hwnd);
-
-        var result = System.Text.Json.JsonSerializer.Deserialize<System.Text.Json.JsonElement>(TestAnsiConsole.Output);
-        Assert.AreEqual("txt-name-1234", result.GetProperty("target").GetString());
-        Assert.AreEqual(4242, result.GetProperty("hwnd").GetInt64());
-    }
-
-    [TestMethod]
-    public async Task SendKeys_TargetNotFound_ReturnsError()
-    {
-        _fakeUia.FindSingleResult = null;
-
-        var command = GetRequiredService<UiSendKeysCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command,
-            ["hello", "-a", "TestApp", "--target", "missing-0000", "--json"]);
-        Assert.AreEqual(1, exitCode);
-        Assert.AreEqual(0, _fakeKeyboard.SendCalls.Count);
-    }
-
-    [TestMethod]
-    public async Task SendKeys_MissingKeys_ReturnsError()
-    {
-        var command = GetRequiredService<UiSendKeysCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command, ["-a", "TestApp", "--json"]);
-        Assert.AreEqual(1, exitCode);
-    }
-
-    [TestMethod]
-    public async Task SendKeys_MissingApp_ReturnsError()
-    {
-        var command = GetRequiredService<UiSendKeysCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command, ["enter", "--json"]);
-        Assert.AreEqual(1, exitCode);
-    }
-
-    [TestMethod]
-    public async Task SendKeys_InvalidVia_ReturnsError()
-    {
-        var command = GetRequiredService<UiSendKeysCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command, ["enter", "-a", "TestApp", "--via", "bogus", "--json"]);
-        Assert.AreEqual(1, exitCode);
-        Assert.AreEqual(0, _fakeKeyboard.SendCalls.Count);
-    }
-
-    [TestMethod]
-    public async Task SendKeys_InvalidKeyToken_ReturnsError()
-    {
-        var command = GetRequiredService<UiSendKeysCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command, ["vk=0xZZ", "-a", "TestApp", "--json"]);
-        Assert.AreEqual(1, exitCode);
-        Assert.AreEqual(0, _fakeKeyboard.SendCalls.Count);
-    }
-
-    [TestMethod]
-    public async Task SendKeys_SystemCombo_ViaSendInput_IsRejected()
-    {
-        // send-input is OS-wide, so a system-reserved combo (win+l) acts on the shell, not just the
-        // target — the command must refuse it and never reach the keyboard transport.
-        var command = GetRequiredService<UiSendKeysCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command,
-            ["win+l", "-a", "TestApp", "--via", "send-input", "--json"]);
-        Assert.AreEqual(1, exitCode);
-        Assert.AreEqual(0, _fakeKeyboard.SendCalls.Count);
-    }
-
-    [TestMethod]
-    public async Task SendKeys_SystemCombo_ViaPostMessage_StillSends()
-    {
-        // post-message is window-scoped (posts straight to the target HWND's queue), so it isn't
-        // OS-wide and is not blocked — an alt+f4 posted to the window just closes that window.
-        var command = GetRequiredService<UiSendKeysCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command,
-            ["alt+f4", "-a", "TestApp", "--via", "post-message", "--json"]);
-        Assert.AreEqual(0, exitCode);
-        Assert.AreEqual(1, _fakeKeyboard.SendCalls.Count);
-    }
-
-    [TestMethod]
-    public void SendKeys_VerbatimOption_DocumentedInDescription()
-    {
-        // The --verbatim flag must be discoverable via --help / cli-schema and explain it types literally.
-        StringAssert.Contains(UiSendKeysCommand.VerbatimOption.Description, "literal");
-    }
-
-    [TestMethod]
-    public async Task SendKeys_Verbatim_TypesEntireArgumentAsLiteralText()
-    {
-        // Without --verbatim this presses Down, Down, Enter; with it, the words are typed verbatim as
-        // a single literal text action.
-        var command = GetRequiredService<UiSendKeysCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command,
-            ["down down enter", "-a", "TestApp", "--verbatim", "--json"]);
-        Assert.AreEqual(0, exitCode);
-
-        Assert.AreEqual(1, _fakeKeyboard.SendCalls.Count);
-        var actions = _fakeKeyboard.SendCalls[0].Actions;
-        Assert.AreEqual(1, actions.Count);
-        Assert.IsInstanceOfType<WinApp.Cli.Helpers.TextInput>(actions[0]);
-        Assert.AreEqual("down down enter", ((WinApp.Cli.Helpers.TextInput)actions[0]).Text);
-
-        var result = System.Text.Json.JsonSerializer.Deserialize<System.Text.Json.JsonElement>(TestAnsiConsole.Output);
-        Assert.AreEqual(1, result.GetProperty("actionCount").GetInt32());
-    }
-
-    [TestMethod]
-    public async Task SendKeys_Verbatim_PreservesExactWhitespace()
-    {
-        // The normal path collapses internal whitespace to a single space; --verbatim keeps it exact.
-        var command = GetRequiredService<UiSendKeysCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command, ["a  b", "-a", "TestApp", "--verbatim"]);
-        Assert.AreEqual(0, exitCode);
-
-        Assert.AreEqual(1, _fakeKeyboard.SendCalls.Count);
-        var actions = _fakeKeyboard.SendCalls[0].Actions;
-        Assert.AreEqual(1, actions.Count);
-        Assert.AreEqual("a  b", ((WinApp.Cli.Helpers.TextInput)actions[0]).Text);
-    }
-
-    [TestMethod]
-    public async Task SendKeys_Verbatim_SystemComboTextViaSendInput_IsSentAsText()
-    {
-        // With --verbatim, "win+l" is literal text (a TextInput), not a chord — the system-combo guard
-        // only inspects key chords, so the text is typed rather than refused even via send-input.
-        var command = GetRequiredService<UiSendKeysCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command,
-            ["win+l", "-a", "TestApp", "--via", "send-input", "--verbatim", "--json"]);
-        Assert.AreEqual(0, exitCode);
-
-        Assert.AreEqual(1, _fakeKeyboard.SendCalls.Count);
-        Assert.IsInstanceOfType<WinApp.Cli.Helpers.TextInput>(_fakeKeyboard.SendCalls[0].Actions[0]);
-    }
-
-    [TestMethod]
-    public async Task SendKeys_Verbatim_MissingKeys_ReturnsError()
-    {
-        // --verbatim still requires a value; an empty keys argument is rejected before injection.
-        var command = GetRequiredService<UiSendKeysCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command, ["-a", "TestApp", "--verbatim", "--json"]);
-        Assert.AreEqual(1, exitCode);
-        Assert.AreEqual(0, _fakeKeyboard.SendCalls.Count);
-    }
-
-    // ---------------------------------------------------------------------
-    // drag (#498) — mouse drag gesture: drag <from> <to>, each a selector
-    // (element center) or app x,y coordinates
-    // ---------------------------------------------------------------------
-
-    [TestMethod]
-    public async Task Drag_RightButton_SetsButton()
-    {
-        _fakeUia.FindSingleResult = new UiElement { Id = "e0", Type = "Image", Selector = "img-canvas-1234", X = 0, Y = 0, Width = 100, Height = 100 };
-
-        var command = GetRequiredService<UiDragCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command, ["img-canvas-1234", "200,200", "-a", "TestApp", "--right", "--json"]);
-        Assert.AreEqual(0, exitCode);
-
-        Assert.AreEqual(1, _fakeMouse.DragCalls.Count);
-        Assert.IsTrue(_fakeMouse.DragCalls[0].RightButton);
-
-        var result = System.Text.Json.JsonSerializer.Deserialize<System.Text.Json.JsonElement>(TestAnsiConsole.Output);
-        Assert.AreEqual("right", result.GetProperty("button").GetString());
-    }
-
-    [TestMethod]
-    public async Task Drag_MissingApp_ReturnsError()
-    {
-        var command = GetRequiredService<UiDragCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command, ["100,100", "200,200", "--json"]);
-        Assert.AreEqual(1, exitCode);
-        Assert.AreEqual(0, _fakeMouse.DragCalls.Count);
-    }
-
-    [TestMethod]
-    public async Task Drag_MissingArgs_ReturnsError()
-    {
-        var command = GetRequiredService<UiDragCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command, ["-a", "TestApp", "--json"]);
-        Assert.AreEqual(1, exitCode);
-        Assert.AreEqual(0, _fakeMouse.DragCalls.Count);
-    }
-
-    // ---- drag <from> <to> where each is a selector (center) or x,y app coords ----
-
-    [TestMethod]
-    public async Task Drag_TwoArg_SelectorToCoordinates_UsesElementCenter()
-    {
-        _fakeUia.FindSingleResult = new UiElement { Id = "e0", Type = "Image", Selector = "img-canvas-1234", X = 50, Y = 60, Width = 200, Height = 200 };
-
-        var command = GetRequiredService<UiDragCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command, ["img-canvas-1234", "300,400", "-a", "TestApp", "--json"]);
-        Assert.AreEqual(0, exitCode);
-
-        var result = System.Text.Json.JsonSerializer.Deserialize<System.Text.Json.JsonElement>(TestAnsiConsole.Output);
-        Assert.AreEqual("img-canvas-1234", result.GetProperty("from").GetString());
-        Assert.AreEqual("300,400", result.GetProperty("to").GetString());
-        Assert.AreEqual(150, result.GetProperty("fromX").GetInt32());  // 50 + 200/2
-        Assert.AreEqual(160, result.GetProperty("fromY").GetInt32());  // 60 + 200/2
-        Assert.AreEqual(300, result.GetProperty("toX").GetInt32());
-        Assert.AreEqual(400, result.GetProperty("toY").GetInt32());
-
-        Assert.AreEqual(1, _fakeMouse.DragCalls.Count);
-        var drag = _fakeMouse.DragCalls[0];
-        Assert.AreEqual(150, drag.FromX);
-        Assert.AreEqual(160, drag.FromY);
-        Assert.AreEqual(300, drag.ToX);
-        Assert.AreEqual(400, drag.ToY);
-    }
-
-    [TestMethod]
-    public async Task Drag_TwoArg_CoordinatesToCoordinates_NoElementLookup()
-    {
-        // Both endpoints are bare coordinates, so no element is resolved (FindSingleResult stays null).
-        _fakeUia.FindSingleResult = null;
-
-        var command = GetRequiredService<UiDragCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command, ["100,200", "300,400", "-a", "TestApp", "--json"]);
-        Assert.AreEqual(0, exitCode);
-
-        var result = System.Text.Json.JsonSerializer.Deserialize<System.Text.Json.JsonElement>(TestAnsiConsole.Output);
-        Assert.AreEqual("100,200", result.GetProperty("from").GetString());
-        Assert.AreEqual("300,400", result.GetProperty("to").GetString());
-        Assert.AreEqual(100, result.GetProperty("fromX").GetInt32());
-        Assert.AreEqual(200, result.GetProperty("fromY").GetInt32());
-        Assert.AreEqual(300, result.GetProperty("toX").GetInt32());
-        Assert.AreEqual(400, result.GetProperty("toY").GetInt32());
-
-        Assert.AreEqual(1, _fakeMouse.DragCalls.Count);
-        var drag = _fakeMouse.DragCalls[0];
-        Assert.AreEqual(100, drag.FromX);
-        Assert.AreEqual(200, drag.FromY);
-        Assert.AreEqual(300, drag.ToX);
-        Assert.AreEqual(400, drag.ToY);
-    }
-
-    [TestMethod]
-    public async Task Drag_TwoArg_SelectorToSelector_UsesBothCenters()
-    {
-        // Distinct from/to elements (via the per-call queue) so reusing one endpoint for both would fail.
-        _fakeUia.FindSingleResults.Enqueue(new UiElement { Id = "e0", Type = "ListItem", Selector = "row-1111", X = 10, Y = 20, Width = 100, Height = 80 });   // center 60,60
-        _fakeUia.FindSingleResults.Enqueue(new UiElement { Id = "e1", Type = "ListItem", Selector = "row-2222", X = 200, Y = 300, Width = 40, Height = 60 });  // center 220,330
-
-        var command = GetRequiredService<UiDragCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command, ["row-1111", "row-2222", "-a", "TestApp", "--json"]);
-        Assert.AreEqual(0, exitCode);
-
-        var result = System.Text.Json.JsonSerializer.Deserialize<System.Text.Json.JsonElement>(TestAnsiConsole.Output);
-        Assert.AreEqual("row-1111", result.GetProperty("from").GetString());
-        Assert.AreEqual("row-2222", result.GetProperty("to").GetString());
-        Assert.AreEqual(60, result.GetProperty("fromX").GetInt32());   // 10 + 100/2
-        Assert.AreEqual(60, result.GetProperty("fromY").GetInt32());   // 20 + 80/2
-        Assert.AreEqual(220, result.GetProperty("toX").GetInt32());    // 200 + 40/2
-        Assert.AreEqual(330, result.GetProperty("toY").GetInt32());    // 300 + 60/2
-
-        Assert.AreEqual(1, _fakeMouse.DragCalls.Count);
-        var drag = _fakeMouse.DragCalls[0];
-        Assert.AreEqual(60, drag.FromX);
-        Assert.AreEqual(60, drag.FromY);
-        Assert.AreEqual(220, drag.ToX);
-        Assert.AreEqual(330, drag.ToY);
-    }
-
-    [TestMethod]
-    public async Task Drag_TwoArg_CoordinatesToSelector_ResolvesOnlyTo()
-    {
-        // from is bare coords (no lookup); to is a selector resolving to its center.
-        _fakeUia.FindSingleResult = new UiElement { Id = "e0", Type = "ListItem", Selector = "row-2222", X = 200, Y = 300, Width = 40, Height = 60 };
-
-        var command = GetRequiredService<UiDragCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command, ["100,200", "row-2222", "-a", "TestApp", "--json"]);
-        Assert.AreEqual(0, exitCode);
-
-        var result = System.Text.Json.JsonSerializer.Deserialize<System.Text.Json.JsonElement>(TestAnsiConsole.Output);
-        Assert.AreEqual("100,200", result.GetProperty("from").GetString());
-        Assert.AreEqual("row-2222", result.GetProperty("to").GetString());
-        Assert.AreEqual(100, result.GetProperty("fromX").GetInt32());
-        Assert.AreEqual(200, result.GetProperty("fromY").GetInt32());
-        Assert.AreEqual(220, result.GetProperty("toX").GetInt32());   // 200 + 40/2
-        Assert.AreEqual(330, result.GetProperty("toY").GetInt32());   // 300 + 60/2
-
-        Assert.AreEqual(1, _fakeMouse.DragCalls.Count);
-    }
-
-    [TestMethod]
-    public async Task Drag_TwoArg_ToSelectorNotFound_ReturnsError()
-    {
-        // from resolves fine; the second (to) selector lookup returns null → error, no drag.
-        _fakeUia.FindSingleResults.Enqueue(new UiElement { Id = "e0", Type = "ListItem", Selector = "row-1111", X = 10, Y = 20, Width = 100, Height = 80 });
-        _fakeUia.FindSingleResults.Enqueue(null);
-
-        var command = GetRequiredService<UiDragCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command, ["row-1111", "missing-9999", "-a", "TestApp", "--json"]);
-        Assert.AreEqual(1, exitCode);
-        Assert.AreEqual(0, _fakeMouse.DragCalls.Count);
-    }
-
-    [TestMethod]
-    public async Task Drag_TwoArg_MissingTo_ReturnsError()
-    {
-        var command = GetRequiredService<UiDragCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command, ["100,200", "-a", "TestApp", "--json"]);
-        Assert.AreEqual(1, exitCode);
-        Assert.AreEqual(0, _fakeMouse.DragCalls.Count);
-    }
-
-    [TestMethod]
-    public async Task Drag_TwoArg_SelectorEndpointNotFound_ReturnsError()
-    {
-        _fakeUia.FindSingleResult = null;
-
-        var command = GetRequiredService<UiDragCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command, ["missing-0000", "300,400", "-a", "TestApp", "--json"]);
-        Assert.AreEqual(1, exitCode);
-        Assert.AreEqual(0, _fakeMouse.DragCalls.Count);
-    }
-
-    [TestMethod]
-    public async Task Drag_TwoArg_ZeroSizeSelector_ReturnsError()
-    {
-        _fakeUia.FindSingleResult = new UiElement { Id = "e0", Type = "Image", Selector = "img-zero-0000", X = 10, Y = 20, Width = 0, Height = 0 };
-
-        var command = GetRequiredService<UiDragCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command, ["img-zero-0000", "300,400", "-a", "TestApp", "--json"]);
-        Assert.AreEqual(1, exitCode);
-        Assert.AreEqual(0, _fakeMouse.DragCalls.Count);
-    }
-
-    // ---- --hold-ms / --dwell-ms: press-and-hold (long-press) and drop-target dwell ----
-
-    [TestMethod]
-    public async Task Drag_HoldMs_FlowsToMouseInput()
-    {
-        // from == to with --hold-ms is a press-and-hold / long-press gesture.
-        _fakeUia.FindSingleResult = new UiElement { Id = "e0", Type = "Button", Selector = "btn-tile-9001", X = 100, Y = 100, Width = 80, Height = 40 };
-
-        var command = GetRequiredService<UiDragCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command, ["btn-tile-9001", "btn-tile-9001", "-a", "TestApp", "--hold-ms", "600", "--json"]);
-        Assert.AreEqual(0, exitCode);
-
-        Assert.AreEqual(1, _fakeMouse.DragCalls.Count);
-        var drag = _fakeMouse.DragCalls[0];
-        Assert.AreEqual(600, drag.HoldMs);
-        Assert.AreEqual(0, drag.DwellMs);
-        Assert.AreEqual(140, drag.FromX); // 100 + 80/2
-        Assert.AreEqual(140, drag.ToX);   // same element → long-press, no movement
-
-        var result = System.Text.Json.JsonSerializer.Deserialize<System.Text.Json.JsonElement>(TestAnsiConsole.Output);
-        Assert.AreEqual(600, result.GetProperty("holdMs").GetInt32());
-        Assert.AreEqual(0, result.GetProperty("dwellMs").GetInt32());
-    }
-
-    [TestMethod]
-    public async Task Drag_DwellMs_FlowsToMouseInput()
-    {
-        // --dwell-ms holds at the destination so a drop target / merge overlay can latch before release.
-        var command = GetRequiredService<UiDragCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command, ["100,100", "300,300", "-a", "TestApp", "--dwell-ms", "350", "--json"]);
-        Assert.AreEqual(0, exitCode);
-
-        Assert.AreEqual(1, _fakeMouse.DragCalls.Count);
-        var drag = _fakeMouse.DragCalls[0];
-        Assert.AreEqual(0, drag.HoldMs);
-        Assert.AreEqual(350, drag.DwellMs);
-
-        var result = System.Text.Json.JsonSerializer.Deserialize<System.Text.Json.JsonElement>(TestAnsiConsole.Output);
-        Assert.AreEqual(350, result.GetProperty("dwellMs").GetInt32());
-    }
-
-    [TestMethod]
-    public async Task Drag_HoldAndDwell_BothFlow()
-    {
-        var command = GetRequiredService<UiDragCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command, ["100,100", "300,300", "-a", "TestApp", "--hold-ms", "120", "--dwell-ms", "200", "--json"]);
-        Assert.AreEqual(0, exitCode);
-
-        var drag = _fakeMouse.DragCalls[0];
-        Assert.AreEqual(120, drag.HoldMs);
-        Assert.AreEqual(200, drag.DwellMs);
-    }
-
-    [TestMethod]
-    public async Task Drag_DefaultHoldAndDwell_AreZero()
-    {
-        var command = GetRequiredService<UiDragCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command, ["100,100", "300,300", "-a", "TestApp", "--json"]);
-        Assert.AreEqual(0, exitCode);
-
-        var drag = _fakeMouse.DragCalls[0];
-        Assert.AreEqual(0, drag.HoldMs);
-        Assert.AreEqual(0, drag.DwellMs);
-    }
-
-    [TestMethod]
-    [DataRow("--hold-ms")]
-    [DataRow("--dwell-ms")]
-    public async Task Drag_NegativeHoldOrDwell_ReturnsError(string option)
-    {
-        var command = GetRequiredService<UiDragCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command, ["100,100", "300,300", "-a", "TestApp", option, "-5", "--json"]);
-        Assert.AreEqual(1, exitCode);
-        Assert.AreEqual(0, _fakeMouse.DragCalls.Count);
-    }
-
-    // ---------------------------------------------------------------------
-    // scroll --wheel (#498) — synthetic mouse-wheel input
-    // ---------------------------------------------------------------------
-
-    [TestMethod]
-    public async Task Scroll_Wheel_SendsWheelDeltaAtElementCenter()
-    {
-        _fakeUia.FindSingleResult = new UiElement { Id = "e0", Type = "List", Selector = "lst-items-1234", X = 50, Y = 60, Width = 120, Height = 40 };
-
-        var command = GetRequiredService<UiScrollCommand>();
-        // --wheel is in notches; -1 notch scales to the -120 WHEEL_DELTA the OS wheel consumes.
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command, ["lst-items-1234", "-a", "TestApp", "--wheel", "-1", "--json"]);
-        Assert.AreEqual(0, exitCode);
-
-        Assert.AreEqual(1, _fakeMouse.ScrollWheelCalls.Count);
-        var wheel = _fakeMouse.ScrollWheelCalls[0];
-        Assert.AreEqual(110, wheel.ScreenX); // 50 + 120/2
-        Assert.AreEqual(80, wheel.ScreenY);  // 60 + 40/2
-        Assert.AreEqual(-120, wheel.Delta);  // -1 notch * 120 (WHEEL_DELTA)
-
-        var result = System.Text.Json.JsonSerializer.Deserialize<System.Text.Json.JsonElement>(TestAnsiConsole.Output);
-        Assert.AreEqual(-1, result.GetProperty("wheel").GetInt32()); // echoes the notch count the caller passed
-    }
-
-    [TestMethod]
-    public async Task Scroll_Wheel_MultipleNotches_ScaleByWheelDelta()
-    {
-        _fakeUia.FindSingleResult = new UiElement { Id = "e0", Type = "List", Selector = "lst-items-1234", X = 0, Y = 0, Width = 100, Height = 100 };
-
-        var command = GetRequiredService<UiScrollCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command, ["lst-items-1234", "-a", "TestApp", "--wheel", "3", "--json"]);
-        Assert.AreEqual(0, exitCode);
-
-        Assert.AreEqual(1, _fakeMouse.ScrollWheelCalls.Count);
-        Assert.AreEqual(360, _fakeMouse.ScrollWheelCalls[0].Delta); // 3 notches * 120
-    }
-
-    [TestMethod]
-    public async Task Scroll_NoDirectionToOrWheel_ReturnsError()
-    {
-        _fakeUia.FindSingleResult = new UiElement { Id = "e0", Type = "List", Selector = "lst-items-1234", X = 0, Y = 0, Width = 100, Height = 100 };
-
-        var command = GetRequiredService<UiScrollCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command, ["lst-items-1234", "-a", "TestApp", "--json"]);
-        Assert.AreEqual(1, exitCode);
-        Assert.AreEqual(0, _fakeMouse.ScrollWheelCalls.Count);
-    }
-
-    [TestMethod]
-    public async Task Scroll_Wheel_ZeroSizeElement_ReturnsError()
-    {
-        _fakeUia.FindSingleResult = new UiElement { Id = "e0", Type = "List", Selector = "lst-tiny-0000", X = 10, Y = 20, Width = 0, Height = 0 };
-
-        var command = GetRequiredService<UiScrollCommand>();
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command, ["lst-tiny-0000", "-a", "TestApp", "--wheel", "-1", "--json"]);
-        Assert.AreEqual(1, exitCode);
-        Assert.AreEqual(0, _fakeMouse.ScrollWheelCalls.Count);
-    }
-
-    [TestMethod]
-    public async Task Scroll_ConflictingModes_ReturnsError()
-    {
-        _fakeUia.FindSingleResult = new UiElement { Id = "e0", Type = "List", Selector = "lst-items-1234", X = 0, Y = 0, Width = 100, Height = 100 };
-
-        var command = GetRequiredService<UiScrollCommand>();
-        // --wheel and --direction are mutually exclusive; specifying both must fail rather than
-        // silently preferring --wheel.
-        var exitCode = await ParseAndInvokeWithCaptureAsync(command, ["lst-items-1234", "-a", "TestApp", "--wheel", "-1", "--direction", "down", "--json"]);
         Assert.AreEqual(1, exitCode);
         Assert.AreEqual(0, _fakeMouse.ScrollWheelCalls.Count);
     }
@@ -1430,4 +889,3 @@ public class UiCommandTests : BaseCommandTests
         StringAssert.Contains(TestAnsiConsole.Output, "\"elementId\":");
     }
 }
-
