+++ src/winapp-CLI/WinApp.Cli/Helpers/RealMouseInput.cs
@@ -15,9 +15,9 @@ internal class RealMouseInput : IMouseInput
     public void Click(int screenX, int screenY, bool doubleClick = false, bool rightClick = false, int settleMs = 50)
         => MouseInput.Click(screenX, screenY, doubleClick, rightClick, settleMs);
 
-    public void Drag(int fromScreenX, int fromScreenY, int toScreenX, int toScreenY, bool rightButton = false, int holdMs = 0, int dwellMs = 0)
-        => MouseInput.Drag(fromScreenX, fromScreenY, toScreenX, toScreenY, rightButton, holdMs, dwellMs);
+    public void Drag(int fromScreenX, int fromScreenY, int toScreenX, int toScreenY, bool rightButton = false, int holdMs = 0, int dwellMs = 0, int settleMs = 50)
+        => MouseInput.Drag(fromScreenX, fromScreenY, toScreenX, toScreenY, rightButton, holdMs, dwellMs, settleMs);
 
-    public void ScrollWheel(int screenX, int screenY, int delta)
-        => MouseInput.ScrollWheel(screenX, screenY, delta);
+    public void ScrollWheel(int screenX, int screenY, int delta, int settleMs = 30)
+        => MouseInput.ScrollWheel(screenX, screenY, delta, settleMs);
 }
