+++ src/winapp-CLI/WinApp.Cli/NativeMethods.txt
@@ -77,6 +77,7 @@ IsIconic
 ShowWindow
 FindWindowEx
 GetWindowThreadProcessId
+GetAncestor
 GetWindowText
 IsWindowVisible
 SetForegroundWindow
